+++ packages/sdk/cryptographic-utils/src/bls.ts
@@ -1,7 +1,7 @@
 // this is an implementation of a subset of BLS12-377
+import { isValidAddress } from '@celo/utils/lib/address'
 import { BLS } from 'bls12377js'
 import { keccak256 } from 'ethereumjs-util'
-import { isValidAddress } from './address'
 const BigInteger = require('bigi')
 const reverse = require('buffer-reverse')