+++ packages/sdk/utils/src/index.ts
@@ -1,12 +1,9 @@
 import * as AddressUtils from './address'
 
-export { AccountUtils } from './account'
 export { AttestationUtils } from './attestations'
-export { CommentEncryptionUtils } from './commentEncryption'
 export * from './contacts'
 export * from './currencies'
 export * from './dappkit'
-export { DataEncryptionKeyUtils } from './dataEncryptionKey'
 export { ECIES } from './ecies'
 export { IstanbulUtils } from './istanbul'
 export { SignatureUtils } from './signatureUtils'